Renamed Proj to Project as requested by PR comment.
