+++ austrakka/components/dashboard/project/__init__.py
@@ -2,22 +2,23 @@ import click
 
 from austrakka.utils.output import table_format_option
 from austrakka.utils.cmd_filter import hide_admin_cmds
-from austrakka.components.dashboard.proj.funcs import add_dashboard
-from austrakka.components.dashboard.proj.funcs import update_dashboard
-from austrakka.components.dashboard.proj.funcs import list_dashboards
-from austrakka.components.dashboard.proj.funcs import rename_dashboard
+from austrakka.components.dashboard.project.funcs import add_dashboard
+from austrakka.components.dashboard.project.funcs import update_dashboard
+from austrakka.components.dashboard.project.funcs import list_dashboards
+from austrakka.components.dashboard.project.funcs import rename_dashboard
 from ....utils.options import *
 
 
+# pylint: disable=duplicate-code
 @click.group()
 @click.pass_context
-def proj(ctx):
+def project(ctx):
     """Commands related to defining a project dashboard, which is a type of dashboard
     for showing in a project UI."""
     ctx.context = ctx.parent.context
 
 
-@proj.command('add', hidden=hide_admin_cmds())
+@project.command('add', hidden=hide_admin_cmds())
 @opt_name(help="Dashboard name. Must be unique.")
 @opt_widget()
 def dashboard_add(name: str, widget_details: [str]):
@@ -25,7 +26,7 @@ def dashboard_add(name: str, widget_details: [str]):
     add_dashboard(name, widget_details)
 
 
-@proj.command('update', hidden=hide_admin_cmds())
+@project.command('update', hidden=hide_admin_cmds())
 @opt_name(help="Dashboard name. Must be unique.")
 @opt_widget()
 @click.argument('dashboard-id', type=int)
@@ -34,14 +35,14 @@ def dashboard_update(dashboard_id: int, name: str, widget_details: [str]):
     update_dashboard(dashboard_id, name, widget_details)
 
 
-@proj.command('list', hidden=hide_admin_cmds())
+@project.command('list', hidden=hide_admin_cmds())
 @table_format_option()
 def dashboards_list(out_format: str):
     """List dashboards available for use in a given project."""
     list_dashboards(out_format)
 
 
-@proj.command('rename', hidden=hide_admin_cmds())
+@project.command('rename', hidden=hide_admin_cmds())
 @click.argument('dashboard-id', type=int)
 @opt_new_name(help="New name of dashoard.", )
 def dashboard_rename(dashboard_id: int, new_name: str):
